ew
ew

+++ FhsingAnything/Assets/1_Scripts/Fishing/FinishLine.cs
@@ -15,8 +15,8 @@ public class FinishLine : MonoBehaviour
     {
         if(isScanned == false && Input.anyKeyDown)
         {
-            GameManager.instance.Stamina -= 1f;
-            Fishing.instance.PlayerHpBar.value = GameManager.instance.Stamina / Fishing.instance.PlayerMaxHp;
+            GameManager.instance.Stamina -= 1;
+            Fishing.instance.PlayerHpBar.value = (float)GameManager.instance.Stamina / (float)Fishing.instance.PlayerMaxHp;
         }       
     }
